refactor(ebpf): remove syscalls main, keep perf_events main around
- syscalls main has been deprecated by SyscallPublisher
- perf_events main is deprecated, but I'll replace it later on
