+++ sikte/src/publishers/mod.rs
@@ -3,4 +3,6 @@ mod perf_events;
 pub mod syscalls;
 
 pub use event_publisher::EventPublisher;
-pub use syscalls::main::syscalls;
+
+// DEPRECATED: we need to migrate this to a Publisher model
+pub use perf_events::main::perf_events;
