+++ sikte/src/publishers/syscalls/mod.rs
@@ -1,4 +1,3 @@
-pub mod main;
 mod publisher;
 mod table;
